Refactor handlers into separate files
