+++ handlers/handlers_test.go
@@ -2,7 +2,6 @@ package handlers
 
 import (
 	"errors"
-	"github.com/ONSdigital/dp-api-clients-go/v2/filter"
 	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
@@ -10,7 +9,6 @@ import (
 	"testing"
 
 	"github.com/ONSdigital/dp-api-clients-go/v2/dataset"
-	"github.com/ONSdigital/dp-api-clients-go/v2/zebedee"
 	"github.com/ONSdigital/dp-frontend-dataset-controller/config"
 	"github.com/golang/mock/gomock"
 	"github.com/gorilla/mux"
@@ -91,359 +89,6 @@ func TestUnitHandlers(t *testing.T) {
 
 			So(w.Code, ShouldEqual, http.StatusInternalServerError)
 		})
-	})
-
-	Convey("test CreateFilterID", t, func() {
-		mockCfg := config.Config{}
-
-		Convey("test CreateFilterID handler, creates a filter id and redirects", func() {
-			mockClient := NewMockFilterClient(mockCtrl)
-			mockClient.EXPECT().CreateBlueprint(ctx, userAuthToken, serviceAuthToken, "", collectionID, "1234", "5678", "2017", []string{"aggregate", "time"}).Return("12345", "testETag", nil)
-
-			mockDatasetClient := NewMockDatasetClient(mockCtrl)
-			dims := dataset.VersionDimensions{
-				Items: []dataset.VersionDimension{
-					{
-						Name: "aggregate",
-					},
-					{
-						Name: "time",
-					},
-				},
-			}
-			mockDatasetClient.EXPECT().GetVersionDimensions(ctx, userAuthToken, serviceAuthToken, collectionID, "1234", "5678", "2017").Return(dims, nil)
-			mockDatasetClient.EXPECT().GetOptions(ctx, userAuthToken, serviceAuthToken, collectionID, "1234", "5678", "2017", "aggregate",
-				&dataset.QueryParams{Offset: 0, Limit: 0}).Return(datasetOptions(0, 0), nil)
-			mockDatasetClient.EXPECT().GetOptions(ctx, userAuthToken, serviceAuthToken, collectionID, "1234", "5678", "2017", "time",
-				&dataset.QueryParams{Offset: 0, Limit: 0}).Return(datasetOptions(0, 0), nil)
-
-			body := strings.NewReader("")
-			w := testResponse(301, body, "/datasets/1234/editions/5678/versions/2017/filter", mockClient, mockDatasetClient, false, mockCfg)
-
-			location := w.Header().Get("Location")
-			So(location, ShouldNotBeEmpty)
-
-			So(location, ShouldEqual, "/filters/12345/dimensions")
-		})
-
-		Convey("test CreateFilterID returns 500 if unable to create a blueprint on filter api", func() {
-			mockClient := NewMockFilterClient(mockCtrl)
-			mockClient.EXPECT().CreateBlueprint(ctx, userAuthToken, serviceAuthToken, "", collectionID, "1234", "5678", "2017", gomock.Any()).Return("", "", errors.New("unable to create filter blueprint"))
-
-			mockDatasetClient := NewMockDatasetClient(mockCtrl)
-			mockDatasetClient.EXPECT().GetVersionDimensions(ctx, userAuthToken, serviceAuthToken, collectionID, "1234", "5678", "2017").Return(dataset.VersionDimensions{}, nil)
-
-			body := strings.NewReader("")
-
-			testResponse(500, body, "/datasets/1234/editions/5678/versions/2017/filter", mockClient, mockDatasetClient, false, mockCfg)
-		})
-	})
-
-	Convey("test CreateFilterFlexID", t, func() {
-		mockCfg := config.Config{EnableCensusPages: true}
-		mockVersions := dataset.VersionsList{
-			Items: []dataset.Version{
-				{}, // deliberately empty
-				{
-					Dimensions: []dataset.VersionDimension{
-						{
-							Name: "aggregate",
-						},
-						{
-							Name: "time",
-						},
-					},
-				},
-			},
-		}
-
-		Convey("test CreateFilterFlexID handler, creates a filter id and redirect includes dimension name", func() {
-			q := dataset.QueryParams{Offset: 0, Limit: 1000}
-			mockDims := []filter.ModelDimension{
-				{
-					Name: "aggregate",
-				},
-				{
-					Name: "time",
-				},
-			}
-			mockClient := NewMockFilterClient(mockCtrl)
-			mockClient.EXPECT().CreateFlexibleBlueprint(ctx, userAuthToken, serviceAuthToken, "", collectionID, "1234", "2021", "1", mockDims, "Example").
-				Return("12345", "testETag", nil)
-
-			mockDatasetClient := NewMockDatasetClient(mockCtrl)
-			mockDatasetClient.EXPECT().GetVersion(ctx, userAuthToken, serviceAuthToken, "", collectionID, "1234", "2021", "1").
-				Return(mockVersions.Items[1], nil)
-			mockDatasetClient.EXPECT().GetOptions(ctx, userAuthToken, serviceAuthToken, collectionID, "1234", "2021", "1", "aggregate", &q).
-				Return(datasetOptions(0, 0), nil)
-			mockDatasetClient.EXPECT().GetOptions(ctx, userAuthToken, serviceAuthToken, collectionID, "1234", "2021", "1", "time", &q).
-				Return(datasetOptions(0, 0), nil)
-			mockDatasetClient.EXPECT().Get(ctx, userAuthToken, serviceAuthToken, collectionID, "1234").
-				Return(dataset.DatasetDetails{IsBasedOn: &dataset.IsBasedOn{ID: "Example"}}, nil)
-
-			body := strings.NewReader("dimension=aggregate")
-			w := testResponse(301, body, "/datasets/1234/editions/2021/versions/1/filter-flex", mockClient, mockDatasetClient, true, mockCfg)
-
-			location := w.Header().Get("Location")
-			So(location, ShouldNotBeEmpty)
-
-			So(location, ShouldEqual, "/filters/12345/dimensions/aggregate")
-		})
-
-		Convey("test post route fails if config is false", func() {
-			mockCfg := config.Config{EnableCensusPages: false}
-			mockDatasetClient := NewMockDatasetClient(mockCtrl)
-			mockFilterClient := NewMockFilterClient(mockCtrl)
-			body := strings.NewReader("")
-
-			testResponse(500, body, "/datasets/1234/editions/2021/versions/1/filter-flex", mockFilterClient, mockDatasetClient, true, mockCfg)
-		})
-
-		Convey("test CreateFilterFlexID returns 500 if unable to create a blueprint on filter api", func() {
-			mockDatasetClient := NewMockDatasetClient(mockCtrl)
-			mockDatasetClient.EXPECT().GetVersion(ctx, userAuthToken, serviceAuthToken, "", collectionID, "1234", "2021", "1").Return(mockVersions.Items[0], nil)
-			mockDatasetClient.EXPECT().Get(ctx, userAuthToken, serviceAuthToken, collectionID, "1234").Return(dataset.DatasetDetails{IsBasedOn: &dataset.IsBasedOn{}}, nil)
-			mockFilterClient := NewMockFilterClient(mockCtrl)
-			mockFilterClient.EXPECT().CreateFlexibleBlueprint(ctx, userAuthToken, serviceAuthToken, "", collectionID, "1234", "2021", "1", gomock.Any(), "").Return("", "", errors.New("unable to create filter blueprint"))
-			body := strings.NewReader("")
-
-			testResponse(500, body, "/datasets/1234/editions/2021/versions/1/filter-flex", mockFilterClient, mockDatasetClient, true, mockCfg)
-		})
-	})
-
-	Convey("test /data endpoint", t, func() {
-
-		Convey("test successful json response", func() {
-			mockZebedeeClient := NewMockZebedeeClient(mockCtrl)
-			mockDatasetClient := NewMockDatasetClient(mockCtrl)
-			mockFilesAPIClient := NewMockFilesAPIClient(mockCtrl)
-			mockZebedeeClient.EXPECT().Get(ctx, "12345", "/data?uri=/data").Return([]byte(`{"some_json":true}`), nil)
-			mockConfig := config.Config{}
-
-			w := httptest.NewRecorder()
-			req, err := http.NewRequest("GET", "/data", nil)
-			So(err, ShouldBeNil)
-			req.AddCookie(&http.Cookie{Name: "access_token", Value: "12345"})
-
-			router := mux.NewRouter()
-			router.Path("/{uri:.*}").HandlerFunc(LegacyLanding(mockZebedeeClient, mockDatasetClient, mockFilesAPIClient, nil, mockConfig))
-
-			router.ServeHTTP(w, req)
-
-			So(w.Body.String(), ShouldEqual, `{"some_json":true}`)
-		})
-
-		Convey("test status 500 returned if zedbedee get returns error", func() {
-			mockZebedeeClient := NewMockZebedeeClient(mockCtrl)
-			mockDatasetClient := NewMockDatasetClient(mockCtrl)
-			mockFilesAPIClient := NewMockFilesAPIClient(mockCtrl)
-			mockZebedeeClient.EXPECT().Get(ctx, userAuthToken, "/data?uri=/data").Return(nil, errors.New("something went wrong with zebedee"))
-			mockConfig := config.Config{}
-
-			w := httptest.NewRecorder()
-			req, err := http.NewRequest("GET", "/data", nil)
-			So(err, ShouldBeNil)
-			router := mux.NewRouter()
-			router.Path("/{uri:.*}").HandlerFunc(LegacyLanding(mockZebedeeClient, mockDatasetClient, mockFilesAPIClient, nil, mockConfig))
-
-			router.ServeHTTP(w, req)
-
-			So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		})
-	})
-
-	Convey("test legacylanding handler with non /data endpoint", t, func() {
-		Convey("test successful data retrieval and rendering", func() {
-			mockZebedeeClient := NewMockZebedeeClient(mockCtrl)
-			mockDatasetClient := NewMockDatasetClient(mockCtrl)
-			mockFilesAPIClient := NewMockFilesAPIClient(mockCtrl)
-			mockConfig := config.Config{}
-			dlp := zebedee.DatasetLandingPage{URI: "https://helloworld.com"}
-			dlp.Datasets = append(dlp.Datasets, zebedee.Related{Title: "A dataset!", URI: "dataset.com"})
-
-			mockZebedeeClient.EXPECT().GetDatasetLandingPage(ctx, userAuthToken, collectionID, locale, "/somelegacypage").Return(dlp, nil)
-			mockZebedeeClient.EXPECT().GetBreadcrumb(ctx, userAuthToken, collectionID, locale, dlp.URI)
-			mockZebedeeClient.EXPECT().GetDataset(ctx, userAuthToken, collectionID, locale, "dataset.com")
-			mockZebedeeClient.EXPECT().GetHomepageContent(ctx, userAuthToken, collectionID, locale, "/")
-
-			mockRend := NewMockRenderClient(mockCtrl)
-			mockRend.EXPECT().NewBasePageModel().Return(coreModel.NewPage(cfg.PatternLibraryAssetsPath, cfg.SiteDomain))
-			mockRend.EXPECT().BuildPage(gomock.Any(), gomock.Any(), "static")
-
-			w := httptest.NewRecorder()
-			req, err := http.NewRequest("GET", "/somelegacypage", nil)
-			So(err, ShouldBeNil)
-
-			router := mux.NewRouter()
-			router.Path("/{uri:.*}").HandlerFunc(LegacyLanding(mockZebedeeClient, mockDatasetClient, mockFilesAPIClient, mockRend, mockConfig))
-
-			router.ServeHTTP(w, req)
-
-			So(w.Code, ShouldEqual, http.StatusOK)
-		})
-
-		Convey("test status 500 returned when zebedee client returns error retrieving landing page", func() {
-			mockZebedeeClient := NewMockZebedeeClient(mockCtrl)
-			mockDatasetClient := NewMockDatasetClient(mockCtrl)
-			mockFilesAPIClient := NewMockFilesAPIClient(mockCtrl)
-			mockConfig := config.Config{}
-			dlp := zebedee.DatasetLandingPage{}
-			mockZebedeeClient.EXPECT().GetDatasetLandingPage(ctx, userAuthToken, collectionID, locale, "/somelegacypage").Return(dlp, errors.New("something went wrong :("))
-
-			w := httptest.NewRecorder()
-			req, err := http.NewRequest("GET", "/somelegacypage", nil)
-			So(err, ShouldBeNil)
-
-			router := mux.NewRouter()
-			router.Path("/{uri:.*}").HandlerFunc(LegacyLanding(mockZebedeeClient, mockDatasetClient, mockFilesAPIClient, nil, mockConfig))
-
-			router.ServeHTTP(w, req)
-			So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		})
-
-		Convey("test status 500 returned when zebedee client returns error retrieving breadcrumb", func() {
-			mockZebedeeClient := NewMockZebedeeClient(mockCtrl)
-			mockDatasetClient := NewMockDatasetClient(mockCtrl)
-			mockFilesAPIClient := NewMockFilesAPIClient(mockCtrl)
-			mockConfig := config.Config{}
-			dlp := zebedee.DatasetLandingPage{URI: "https://helloworld.com"}
-			mockZebedeeClient.EXPECT().GetDatasetLandingPage(ctx, userAuthToken, collectionID, locale, "/somelegacypage").Return(dlp, nil)
-			mockZebedeeClient.EXPECT().GetBreadcrumb(ctx, userAuthToken, collectionID, locale, dlp.URI).Return(nil, errors.New("something went wrong"))
-
-			w := httptest.NewRecorder()
-			req, err := http.NewRequest("GET", "/somelegacypage", nil)
-			So(err, ShouldBeNil)
-
-			router := mux.NewRouter()
-			router.Path("/{uri:.*}").HandlerFunc(LegacyLanding(mockZebedeeClient, mockDatasetClient, mockFilesAPIClient, nil, mockConfig))
-
-			router.ServeHTTP(w, req)
-			So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		})
-	})
-
-	Convey("test filterable landing page", t, func() {
-
-		Convey("test filterable landing page is successful, when it receives good dataset api responses", func() {
-			mockZebedeeClient := NewMockZebedeeClient(mockCtrl)
-			mockClient := NewMockDatasetClient(mockCtrl)
-			mockConfig := config.Config{}
-			mockClient.EXPECT().Get(ctx, userAuthToken, serviceAuthToken, collectionID, "12345").Return(dataset.DatasetDetails{Contacts: &[]dataset.Contact{{Name: "Matt"}}, URI: "/economy/grossdomesticproduct/datasets/gdpjanuary2018", Links: dataset.Links{LatestVersion: dataset.Link{URL: "/datasets/1234/editions/5678/versions/2017"}}}, nil)
-			versions := dataset.VersionsList{
-				Items: []dataset.Version{
-					{
-						Links:       dataset.Links{Self: dataset.Link{URL: "/datasets/12345/editions/2016/versions/1"}},
-						ReleaseDate: "02-01-2005",
-					},
-				},
-			}
-			mockClient.EXPECT().GetVersions(ctx, userAuthToken, serviceAuthToken, collectionID, "", "12345", "5678", &dataset.QueryParams{Offset: 0, Limit: 1000}).Return(versions, nil)
-			mockClient.EXPECT().GetVersion(ctx, userAuthToken, serviceAuthToken, collectionID, "", "12345", "5678", "2017").Return(versions.Items[0], nil)
-			dims := dataset.VersionDimensions{
-				Items: []dataset.VersionDimension{
-					{
-						Name: "aggregate",
-					},
-				},
-			}
-			mockClient.EXPECT().GetVersionDimensions(ctx, userAuthToken, serviceAuthToken, collectionID, "12345", "5678", "2017").Return(dims, nil)
-			mockClient.EXPECT().GetOptions(ctx, userAuthToken, serviceAuthToken, collectionID, "12345", "5678", "2017", "aggregate",
-				&dataset.QueryParams{Offset: 0, Limit: numOptsSummary}).Return(datasetOptions(0, numOptsSummary), nil)
-			mockClient.EXPECT().GetVersionMetadata(ctx, userAuthToken, serviceAuthToken, collectionID, "12345", "5678", "2017")
-			mockClient.EXPECT().GetOptions(ctx, userAuthToken, serviceAuthToken, collectionID, "12345", "5678", "2017", "aggregate",
-				&dataset.QueryParams{Offset: 0, Limit: maxMetadataOptions}).Return(datasetOptions(0, maxMetadataOptions), nil)
-			mockZebedeeClient.EXPECT().GetBreadcrumb(ctx, userAuthToken, collectionID, locale, "")
-			mockZebedeeClient.EXPECT().GetHomepageContent(ctx, userAuthToken, collectionID, locale, "/")
-
-			mockRend := NewMockRenderClient(mockCtrl)
-			mockRend.EXPECT().NewBasePageModel().Return(coreModel.NewPage(cfg.PatternLibraryAssetsPath, cfg.SiteDomain))
-			mockRend.EXPECT().BuildPage(gomock.Any(), gomock.Any(), "filterable")
-
-			w := httptest.NewRecorder()
-			req := httptest.NewRequest("GET", "/datasets/12345", nil)
-
-			router := mux.NewRouter()
-			router.HandleFunc("/datasets/{datasetID}", FilterableLanding(mockClient, mockRend, mockZebedeeClient, mockConfig, ""))
-
-			router.ServeHTTP(w, req)
-
-			So(w.Code, ShouldEqual, http.StatusOK)
-		})
-
-		Convey("test filterableLanding returns 302 and redirects to the correct url for edition level requests without version", func() {
-			mockZebedeeClient := NewMockZebedeeClient(mockCtrl)
-			mockClient := NewMockDatasetClient(mockCtrl)
-			mockConfig := config.Config{}
-			mockZebedeeClient.EXPECT().GetHomepageContent(ctx, userAuthToken, collectionID, locale, "/")
-			mockClient.EXPECT().Get(ctx, userAuthToken, serviceAuthToken, collectionID, "12345").Return(dataset.DatasetDetails{Contacts: &[]dataset.Contact{{Name: "Matt"}}, URI: "/economy/grossdomesticproduct/datasets/gdpjanuary2018", Links: dataset.Links{LatestVersion: dataset.Link{URL: "/datasets/1234/editions/5678/versions/2017"}}}, nil)
-			versions := dataset.VersionsList{
-				Items: []dataset.Version{
-					{
-						Links:       dataset.Links{Self: dataset.Link{URL: "/datasets/12345/editions/2016/versions/1"}},
-						ReleaseDate: "02-01-2005",
-					},
-				},
-			}
-			mockClient.EXPECT().GetVersions(ctx, userAuthToken, serviceAuthToken, collectionID, "", "12345", "5678", &dataset.QueryParams{Offset: 0, Limit: 1000}).Return(versions, nil)
-
-			mockRend := NewMockRenderClient(mockCtrl)
-
-			w := httptest.NewRecorder()
-			req := httptest.NewRequest("GET", "/datasets/12345/editions/5678", nil)
-
-			router := mux.NewRouter()
-			router.HandleFunc("/datasets/{datasetID}/editions/{editionID}", FilterableLanding(mockClient, mockRend, mockZebedeeClient, mockConfig, ""))
-
-			router.ServeHTTP(w, req)
-
-			So(w.Code, ShouldEqual, http.StatusFound)
-			So(w.Body.String(), ShouldEqual, "<a href=\"/datasets/12345/editions/5678/versions/1\">Found</a>.\n\n")
-		})
-
-		Convey("test filterableLanding returns 500 if client Get() returns an error", func() {
-			mockZebedeeClient := NewMockZebedeeClient(mockCtrl)
-			mockClient := NewMockDatasetClient(mockCtrl)
-			mockConfig := config.Config{}
-			mockClient.EXPECT().Get(ctx, userAuthToken, serviceAuthToken, collectionID, "12345").Return(dataset.DatasetDetails{}, errors.New("sorry"))
-
-			w := httptest.NewRecorder()
-			req := httptest.NewRequest("GET", "/datasets/12345", nil)
-
-			router := mux.NewRouter()
-			router.HandleFunc("/datasets/{datasetID}", FilterableLanding(mockClient, nil, mockZebedeeClient, mockConfig, "/v1"))
-
-			router.ServeHTTP(w, req)
-
-			So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		})
-
-		Convey("test filterableLanding returns 500 if client GetVersions() returns error", func() {
-			mockZebedeeClient := NewMockZebedeeClient(mockCtrl)
-			mockClient := NewMockDatasetClient(mockCtrl)
-			mockConfig := config.Config{}
-			mockClient.EXPECT().Get(ctx, userAuthToken, serviceAuthToken, collectionID, "12345").Return(dataset.DatasetDetails{}, nil)
-			versions := dataset.VersionsList{
-				Items: []dataset.Version{
-					{
-						Links:       dataset.Links{Self: dataset.Link{URL: "/datasets/12345/editions/2016/versions/1"}},
-						ReleaseDate: "02-01-2005",
-					},
-				},
-			}
-			mockClient.EXPECT().GetVersions(ctx, userAuthToken, serviceAuthToken, collectionID, "", "12345", "5678", &dataset.QueryParams{Offset: 0, Limit: 1000}).Return(versions, errors.New("sorry"))
-			mockZebedeeClient.EXPECT().GetHomepageContent(ctx, userAuthToken, collectionID, locale, "/")
-
-			w := httptest.NewRecorder()
-			req := httptest.NewRequest("GET", "/datasets/12345/editions/5678", nil)
-
-			router := mux.NewRouter()
-			router.HandleFunc("/datasets/{datasetID}/editions/{editionID}", FilterableLanding(mockClient, nil, mockZebedeeClient, mockConfig, "/v1"))
-
-			router.ServeHTTP(w, req)
-
-			So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		})
-
 	})
 
 	Convey("test census landing page", t, func() {
@@ -729,47 +374,6 @@ func TestUnitHandlers(t *testing.T) {
 		})
 	})
 
-	Convey("test versions list", t, func() {
-		Convey("test versions list returns 200 when rendered succesfully", func() {
-			mockClient := NewMockDatasetClient(mockCtrl)
-			mockZebedeeClient := NewMockZebedeeClient(mockCtrl)
-			mockZebedeeClient.EXPECT().GetHomepageContent(ctx, userAuthToken, collectionID, locale, "/")
-			mockConfig := config.Config{}
-			mockClient.EXPECT().Get(ctx, userAuthToken, serviceAuthToken, collectionID, "12345").Return(dataset.DatasetDetails{}, nil)
-			mockClient.EXPECT().GetVersions(ctx, userAuthToken, serviceAuthToken, collectionID, "", "12345", "2017", &dataset.QueryParams{Offset: 0, Limit: 1000}).Return(dataset.VersionsList{Items: []dataset.Version{}}, nil)
-			mockClient.EXPECT().GetEdition(ctx, userAuthToken, serviceAuthToken, collectionID, "12345", "2017").Return(dataset.Edition{}, nil)
-
-			mockRend := NewMockRenderClient(mockCtrl)
-			mockRend.EXPECT().NewBasePageModel().Return(coreModel.NewPage(cfg.PatternLibraryAssetsPath, cfg.SiteDomain))
-			mockRend.EXPECT().BuildPage(gomock.Any(), gomock.Any(), "version-list")
-
-			w := httptest.NewRecorder()
-			req := httptest.NewRequest("GET", "/datasets/12345/editions/2017/versions", nil)
-
-			router := mux.NewRouter()
-			router.HandleFunc("/datasets/{datasetID}/editions/{edition}/versions", VersionsList(mockClient, mockZebedeeClient, mockRend, mockConfig))
-
-			router.ServeHTTP(w, req)
-
-			So(w.Code, ShouldEqual, http.StatusOK)
-		})
-
-		Convey("test versions list returns status 500 when dataset client returns an error", func() {
-			mockClient := NewMockDatasetClient(mockCtrl)
-			mockConfig := config.Config{}
-			mockClient.EXPECT().Get(ctx, userAuthToken, serviceAuthToken, collectionID, "12345").Return(dataset.DatasetDetails{}, errors.New("dataset client error"))
-
-			w := httptest.NewRecorder()
-			req := httptest.NewRequest("GET", "/datasets/12345/editions/2017/versions", nil)
-
-			router := mux.NewRouter()
-			router.HandleFunc("/datasets/{datasetID}/editions/{edition}/versions", VersionsList(mockClient, nil, nil, mockConfig))
-
-			router.ServeHTTP(w, req)
-
-			So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		})
-	})
 }
 
 func testResponse(code int, body *strings.Reader, url string, fc FilterClient, dc DatasetClient, filterFlexRoute bool, cfg config.Config) *httptest.ResponseRecorder {
